4.anthropic/3.mcp/4.langchain_bridge/requirements.txt
asyncio
python-dotenv
openai
mcp
langchain
langchain-openai
langchain-core
langchainhub
langgraph
